+++ src/js/features/deck-builder-features/edhrec-suggestions/make-edhrec-button.js
@@ -2,14 +2,11 @@ import bus from 'framebus'
 import {
   BUS_EVENTS as events
 } from 'Constants'
-import mutation from 'Lib/mutation'
 import createElement from 'Lib/create-element'
 import DeckSectionChooser from 'Ui/deck-section-chooser'
 import AddCardElement from 'Ui/add-card-element'
 import Drawer from 'Ui/drawer'
 import scryfall from 'Lib/scryfall'
-import deckParser from 'Lib/deck-parser'
-import iframe from 'Lib/iframe'
 import {
   EDHREC_SYMBOL
 } from 'Svg'
@@ -27,15 +24,7 @@ const TYPES_WITH_IRREGULAR_PLURALS = {
   Sorcery: 'Sorceries'
 }
 
-export default async function makeEDHRecButton () {
-  await iframe.create({
-    // does not matter where on edhrec we open the page
-    // just need to be on the edhrec domain to access
-    // the recs JSON endpoint
-    src: 'https://edhrec.com/404',
-    id: 'edhrec-suggestions-iframe'
-  })
-
+export default function makeEDHRecButton () {
   const button = createElement(`<button
     id="edhrec-suggestions"
     aria-label="EDHRec Suggestions"
@@ -47,24 +36,7 @@ export default async function makeEDHRecButton () {
 </button>`).firstChild
   createDrawer(button)
 
-  const initialCommanders = await getInitialCommanderList()
-  await setDisabledState(button, initialCommanders)
-
-  updateButtonStateOnCommanderChange(button, initialCommanders)
-
   return button
-}
-
-async function setDisabledState (button, commanders) {
-  const allLegal = await deckParser.hasLegalCommanders(commanders)
-
-  // TODO: should we mark it as enabled if at least one commander
-  // is a legal commander, and then only send the legal ones to edhrec?
-  if (allLegal) {
-    button.removeAttribute('disabled')
-  } else {
-    button.setAttribute('disabled', 'disabled')
-  }
 }
 
 function createDrawer (button) {
@@ -105,56 +77,6 @@ function createDrawer (button) {
   })
 
   return drawer
-}
-
-async function getInitialCommanderList () {
-  const initialDeck = await scryfall.getDeck()
-
-  return initialDeck.entries.commanders
-    .filter(filterOutInvalidCards)
-    .map(getCardName)
-    .sort()
-}
-
-function updateButtonStateOnCommanderChange (button, commanders) {
-  mutation.change('.deckbuilder-editor-inner .deckbuilder-column .deckbuilder-section', async (el) => {
-    const title = el.querySelector('.deckbuilder-section-title')
-
-    if (title.innerText.toLowerCase().indexOf('commander') === -1) {
-      // only run mutation on commander column
-      return
-    }
-
-    const commanderList = Array.from(el.querySelectorAll('ul .deckbuilder-entry')).reduce((all, entry) => {
-      // if the select options have more than 2 disabled, this
-      // indicates that the card lookup has not completed, so
-      // we ignore this value
-      const cardLookupNotComplete = entry.querySelectorAll('.deckbuilder-entry-menu-select option[disabled]').length > 2
-
-      if (cardLookupNotComplete) {
-        return all
-      }
-
-      const input = entry.querySelector('.deckbuilder-entry-input')
-      const parts = input.value.trim().match(/^(\d+ )(.*)/)
-      if (!parts) {
-        return all
-      }
-      const name = parts[2]
-
-      all.push(name)
-
-      return all
-    }, [])
-
-    commanderList.sort()
-
-    // hack to determine if the arrays are equal
-    if (commanderList.join('|') !== commanders.join('|')) {
-      commanders = commanderList
-      await setDisabledState(button, commanders)
-    }
-  })
 }
 
 function getCardsInDeck (entries) {