+++ test/features/edhrec-suggestions/make-edhrec-button.test.js
@@ -5,7 +5,6 @@ import wait from 'Lib/wait'
 import Drawer from 'Ui/drawer'
 import mutation from 'Lib/mutation'
 import scryfall from 'Lib/scryfall'
-import iframe from 'Lib/iframe'
 
 describe('makeEDHRecButton', function () {
   beforeEach(function () {
@@ -17,7 +16,6 @@ describe('makeEDHRecButton', function () {
         commanders: []
       }
     })
-    jest.spyOn(deckParser, 'hasLegalCommanders').mockResolvedValue(true)
     jest.spyOn(deckParser, 'getSections').mockReturnValue([
       'commanders',
       'lands',
@@ -30,47 +28,20 @@ describe('makeEDHRecButton', function () {
     deckbuilderElement.id = 'deckbuilder'
     document.body.appendChild(deckbuilderElement)
 
-    jest.spyOn(iframe, 'create').mockResolvedValue()
     // jest doesn't know about the scrollTo method on elements
     jest.spyOn(Drawer.prototype, 'scrollTo').mockImplementation()
   })
 
-  it('makes a button', async function () {
-    const btn = await makeEDHRecButton()
+  it('makes a button', function () {
+    const btn = makeEDHRecButton()
 
     expect(btn.tagName).toBe('BUTTON')
-  })
-
-  it('sets button to disabled when requested deck does not have legal commanders', async function () {
-    deckParser.hasLegalCommanders.mockResolvedValue(false)
-
-    const btn = await makeEDHRecButton()
-
-    expect(btn.getAttribute('disabled')).toBe('disabled')
-  })
-
-  it('sets button to not disabled when requested has legal commanders', async function () {
-    deckParser.hasLegalCommanders.mockResolvedValue(true)
-
-    const btn = await makeEDHRecButton()
-
-    expect(btn.getAttribute('disabled')).toBeFalsy()
   })
 
   it('adds an edhrec drawer to page', async function () {
     await makeEDHRecButton()
 
     expect(document.querySelector('#edhrec-drawer')).not.toBeFalsy()
-  })
-
-  it('adds an edhrec iframe to page', async function () {
-    await makeEDHRecButton()
-
-    expect(iframe.create).toBeCalledTimes(1)
-    expect(iframe.create).toBeCalledWith({
-      src: 'https://edhrec.com/404',
-      id: 'edhrec-suggestions-iframe'
-    })
   })
 
   it('cleans up deck after drawer is closed', async function () {
@@ -101,9 +72,17 @@ describe('makeEDHRecButton', function () {
   })
 
   describe('when clicked', function () {
-    let fakeDeck, fakeEDHRecResponse
+    let fakeDeck, fakeEDHRecResponse, click
 
     beforeEach(async function () {
+      click = (btn) => {
+        // https://stackoverflow.com/a/2706236/2601552
+        const evObj = document.createEvent('Events')
+
+        evObj.initEvent('click', true, false)
+        btn.dispatchEvent(evObj)
+      }
+
       fakeDeck = {
         entries: {
           commanders: [
@@ -248,13 +227,13 @@ describe('makeEDHRecButton', function () {
       await wait()
     })
 
-    it('opens the drawer', async function () {
+    it('opens the drawer', function () {
       bus.emit.mockImplementation()
-      const btn = await makeEDHRecButton()
+      const btn = makeEDHRecButton()
 
       jest.spyOn(Drawer.prototype, 'open')
 
-      btn.click()
+      click(btn)
 
       expect(Drawer.prototype.open).toBeCalledTimes(1)
     })
@@ -263,7 +242,7 @@ describe('makeEDHRecButton', function () {
       bus.emit.mockImplementation()
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -302,7 +281,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -339,7 +318,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -367,7 +346,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -388,7 +367,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -404,7 +383,7 @@ describe('makeEDHRecButton', function () {
 
       jest.spyOn(Drawer.prototype, 'setContent')
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -430,7 +409,7 @@ describe('makeEDHRecButton', function () {
         type_line: 'Instant'
       })
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -457,7 +436,7 @@ describe('makeEDHRecButton', function () {
         type_line: 'Instant'
       })
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -572,7 +551,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -615,7 +594,7 @@ describe('makeEDHRecButton', function () {
 
       const btn = await makeEDHRecButton()
 
-      btn.click()
+      click(btn)
 
       await wait()
 
@@ -625,235 +604,6 @@ describe('makeEDHRecButton', function () {
       expect(sections[0].querySelector('h3').innerHTML).toBe('Instants')
       expect(sections[1].querySelector('h3').innerHTML).toBe('Artifacts')
       expect(sections[2].querySelector('h3').innerHTML).toBe('Lands')
-    })
-  })
-
-  describe('when commander list changes', function () {
-    let fakeDeck, commanderSection
-
-    function addEntry (options = {}) {
-      const li = document.createElement('li')
-      li.classList.add('deckbuilder-entry')
-      li.innerHTML = `
-        <select class="deckbuilder-entry-menu-select">
-          <option id="option-1">1</option>
-          <option id="option-2">2</option>
-          <option id="option-3">3</option>
-          <option id="option-4">4</option>
-          <option id="option-5">5</option>
-        </select>
-        <textarea class="deckbuilder-entry-input"></textarea>
-      `
-      if (options.value) {
-        li.querySelector('textarea').value = options.value
-      }
-      if (options.disabled) {
-        Array.from(li.querySelectorAll('select option'))
-          .forEach(el => el.setAttribute('disabled', 'disabled'))
-      }
-      commanderSection.querySelector('ul').appendChild(li)
-    }
-
-    beforeEach(async function () {
-      fakeDeck = {
-        entries: {
-          commanders: [{
-            card_digest: {
-              name: 'Arjun, the Shifting Flame'
-            }
-          }],
-          lands: [],
-          nonlands: []
-        }
-      }
-      commanderSection = document.createElement('div')
-      commanderSection.innerHTML = `
-        <div class="deckbuilder-section-title"></div>
-        <ul></ul>
-      `
-      commanderSection.querySelector('.deckbuilder-section-title').innerText = 'Commander(s)'
-
-      scryfall.getDeck.mockResolvedValue(fakeDeck)
-      document.body.appendChild(commanderSection)
-    })
-
-    it('enables the button when all entries are commanders', async function () {
-      fakeDeck.entries.commanders = []
-      addEntry()
-      deckParser.hasLegalCommanders.mockResolvedValue(false)
-
-      const btn = await makeEDHRecButton()
-      expect(btn.getAttribute('disabled')).toBeTruthy()
-
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      deckParser.hasLegalCommanders.mockResolvedValue(true)
-      commanderSection.querySelector('textarea').value = '1 Arjun, the Shifting Flame'
-
-      await changeHandler(commanderSection)
-
-      expect(btn.getAttribute('disabled')).toBeFalsy()
-    })
-
-    it('enables the button entry has gone from illegal state to legal state', async function () {
-      fakeDeck.entries.commanders = [{
-        card_digest: {
-          name: 'Food Chain'
-        }
-      }]
-      addEntry({
-        value: '1 Food Chain'
-      })
-      deckParser.hasLegalCommanders.mockResolvedValue(false)
-
-      const btn = await makeEDHRecButton()
-      expect(btn.getAttribute('disabled')).toBeTruthy()
-
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      deckParser.hasLegalCommanders.mockResolvedValue(true)
-      commanderSection.querySelector('textarea').value = '1 Arjun, the Shifting Flame'
-
-      await changeHandler(commanderSection)
-
-      expect(btn.getAttribute('disabled')).toBeFalsy()
-    })
-
-    it('disables the button when the values have changed at least one entry is not a commander', async function () {
-      addEntry({
-        value: '1 Arjun, the Shifting Flame'
-      })
-
-      const btn = await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-      addEntry({
-        value: '1 Rhystic Study'
-      })
-
-      deckParser.hasLegalCommanders.mockResolvedValue(false)
-
-      expect(btn.getAttribute('disabled')).toBeFalsy()
-
-      await changeHandler(commanderSection)
-
-      expect(btn.getAttribute('disabled')).toBeTruthy()
-    })
-
-    it('ignores blank entries', async function () {
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      addEntry()
-      addEntry()
-      addEntry({
-        value: '1 Arjun, the Shifting Flame'
-      })
-      addEntry()
-      addEntry()
-
-      await changeHandler(commanderSection)
-
-      expect(deckParser.hasLegalCommanders).toBeCalledTimes(1)
-      expect(deckParser.hasLegalCommanders).toBeCalledWith([
-        'Arjun, the Shifting Flame'
-      ])
-    })
-
-    it('ignores entries that have not finished lookup', async function () {
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      addEntry({
-        value: '1 Sidar Kondo of Jamuraa'
-      })
-      addEntry({
-        value: '1 Tana the',
-        disabled: true
-      })
-
-      deckParser.hasLegalCommanders.mockClear()
-      await changeHandler(commanderSection)
-
-      expect(deckParser.hasLegalCommanders).toBeCalledTimes(1)
-      expect(deckParser.hasLegalCommanders).toBeCalledWith([
-        'Sidar Kondo of Jamuraa'
-      ])
-    })
-
-    it('ignores entries that do not match deck pattern', async function () {
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      addEntry({
-        value: '1 Sidar Kondo of Jamuraa'
-      })
-      addEntry({
-        value: 'Tana the'
-      })
-
-      deckParser.hasLegalCommanders.mockClear()
-      await changeHandler(commanderSection)
-
-      expect(deckParser.hasLegalCommanders).toBeCalledTimes(1)
-      expect(deckParser.hasLegalCommanders).toBeCalledWith([
-        'Sidar Kondo of Jamuraa'
-      ])
-    })
-
-    it('does not check legality of commanders whe commander list is unchanged', async function () {
-      addEntry({
-        value: '1 Arjun, the Shifting Flame'
-      })
-
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      deckParser.hasLegalCommanders.mockClear()
-      await changeHandler(commanderSection)
-
-      expect(deckParser.hasLegalCommanders).not.toBeCalled()
-    })
-
-    it('does not check legality of commanders whe commander list is unchanged but in a different order', async function () {
-      fakeDeck.entries.commanders = [{
-        card_digest: {
-          name: 'Sidar Kondo of Jamuraa'
-        }
-      }, {
-        card_digest: {
-          name: 'Tana the Bloodsower'
-        }
-      }]
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      deckParser.hasLegalCommanders.mockClear()
-      addEntry({
-        value: '1 Tana the Bloodsower'
-      })
-      addEntry({
-        value: '1 Sidar Kondo of Jamuraa'
-      })
-
-      await changeHandler(commanderSection)
-
-      expect(deckParser.hasLegalCommanders).not.toBeCalled()
-    })
-
-    it('does not check commanders when section is not the commander section', async function () {
-      const title = document.createElement('div')
-      title.innerText = 'Lands'
-      const fakeEl = {
-        querySelector: jest.fn().mockReturnValue(title),
-        querySelectorAll: jest.fn()
-      }
-
-      await makeEDHRecButton()
-      const changeHandler = mutation.change.mock.calls[0][1]
-
-      await changeHandler(fakeEl)
-
-      expect(fakeEl.querySelectorAll).not.toBeCalled()
     })
   })
 })